+++ tests/market/test_instruction_layouts.py
@@ -2,7 +2,7 @@
 from solana.publickey import PublicKey
 
 from pyserum._layouts.instructions import _VERSION, INSTRUCTIONS_LAYOUT, InstructionType
-from pyserum.enums import OrderType, Side
+from pyserum.market.enums import OrderType, Side
 
 
 def assert_parsed_layout(instruction_type, args, raw_bytes):
